Improve error handling and response in UpdaterService
Improved the handling of library updates in UpdaterService by checking for the existence of a folder before attempting to remove it. Exceptions during the removal are now also caught and processed. Moreover, the response status has been changed to "same" instead of "ok" when the library is already up-to-date to provide a more precise status.